single file component: CardResourceCounter, CardRequirementsComponent, CardRenderSymbolComponent, and CardRenderTileComponent (#3466)
