+++ src/components/card/CardProductionBoxComponent.vue
@@ -14,7 +14,7 @@
 
 import Vue from 'vue';
 import CardRenderItemComponent from './CardRenderItemComponent.vue';
-import {CardRenderSymbolComponent} from './CardRenderSymbolComponent';
+import CardRenderSymbolComponent from './CardRenderSymbolComponent.vue';
 import {CardRenderSymbol} from '../../cards/render/CardRenderSymbol';
 import {CardRenderItem} from '../../cards/render/CardRenderItem';
