+++ src/components/card/CardRenderCorpBoxComponent.vue
@@ -14,7 +14,7 @@
 import Vue from 'vue';
 import CardRenderItemComponent from './CardRenderItemComponent.vue';
 import CardRenderEffectBoxComponent from './CardRenderEffectBoxComponent.vue';
-import {CardRenderSymbolComponent} from './CardRenderSymbolComponent';
+import CardRenderSymbolComponent from './CardRenderSymbolComponent.vue';
 import {CardRenderEffect} from '../../cards/render/CardRenderer';
 import {CardRenderItem} from '../../cards/render/CardRenderItem';
 import {CardRenderSymbol} from '../../cards/render/CardRenderSymbol';
